backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
drf-nested-routers==0.93.5
djoser==2.2.2
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.27.1